requirements.txt
datasets==2.18.0
matplotlib==3.7.5
numpy==1.24.4
openai==1.40.8
pandas==2.0.3
peft==0.10.0
python-dotenv==1.0.1
rouge_score==0.1.2
torch==2.2.2
tqdm==4.66.2
transformers==4.41.2
trl==0.9.6
wandb==0.17.7
